Use minimal mistakes from gem instead of forking
Our forked version of MM was a year old and didn't build anymore. This
removes all the theme files from the repo and instead installs MM from a
rubygem, which will keep it up to date as successive container versions
are built. We shouldn't have to update MM or depenencies anymore, and
things should continue building.

Bonus: there are many fewer files to maintain.

Potential drawback: our edits to the theme layouts are gone, but I'm not
    really noticing that they're gone, anyway. I think staying up to date
    is a bigger win.